+++ babel.config.js
@@ -1,5 +1,5 @@
 // babel.config.js
 module.exports = {
-  presets: [['@babel/preset-env', { targets: { node: 'current' }, modules: false }], '@babel/preset-react'],
+  presets: ['@babel/preset-env', '@babel/preset-react'],
   plugins: ['react-hot-loader/babel'],
 };
