+++ package.json
@@ -37,6 +37,7 @@
     "react-test-renderer": "^17.0.1"
   },
   "devDependencies": {
+    "@babel/cli": "7.11.6",
     "@babel/core": "^7.12.10",
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-react": "^7.12.10",
@@ -63,8 +64,5 @@
     "webpack": "^5.13.0",
     "webpack-cli": "^4.3.1",
     "webpack-dev-server": "^3.11.1"
-  },
-  "jest": {
-    "testEnvironment": "node"
   }
 }
